+++ Stage1/perception/sensors.py
@@ -1,4 +1,4 @@
-from environment.env_1d import *
+from environment.enrvironment2d import *
 from functools import reduce
